nom_supreme
Module error

An error type, ErrorTree, designed to retain much more useful information about parse failures than the built-in nom error types. Requires the error feature to be enabled.

Enums

BaseErrorKind

These are the different specific things that can go wrong at a particular location during a nom parse. Many of these are collected into an ErrorTree.

Expectation

Enum for generic things that can be expected by nom parsers

GenericErrorTree

Generic version of ErrorTree, which allows for arbitrary Tag, Context, and ExternalError types. See ErrorTree for more extensive docs and examples.

StackContext

Context that can appear in a stack, above a base ErrorTree. Stack contexts are attached by parser combinators to errors from their subparsers during stack unwinding.

Type Aliases

ErrorTree

A comprehensive tree of nom errors describing a parse failure.