nom_supreme::error
Enum GenericErrorTree

pub enum GenericErrorTree<Location, Tag, Context, ExternalError> {
    Base {
        location: Location,
        kind: BaseErrorKind<Tag, ExternalError>,
    },
    Stack {
        base: Box<Self>,
        contexts: Vec<(Location, StackContext<Context>)>,
    },
    Alt(Vec<Self>),
}

Generic version of ErrorTree, which allows for arbitrary Tag, Context, and ExternalError types. See ErrorTree for more extensive docs and examples.

Tag is typically something like &'static str or &'static [u8].

Context is typically &'static str, and it must be in order to interoperate with nomโs context combinator. nom-supreme provides a more generic .context combinator that allows for any kind of context to be attached to an error.

Error can usually be nothing, as itโs unusual for nom parsers to require external errors. Box<dyn Error + Send + Sync + 'static> is a common catch-all.

Variants

Base

A specific error event at a specific location. Often this will indicate that something like a tag or character was expected at that location.

Fields

location: Location

The location of this error in the input

kind: BaseErrorKind<Tag, ExternalError>

The specific error that occurred

Stack

A stack indicates a chain of error contexts was provided. The stack should be read โbackwardsโ; that is, errors earlier in the Vec occurred โsoonerโ (deeper in the call stack).

Fields

base: Box<Self>

The original error

contexts: Vec<(Location, StackContext<Context>)>

The stack of contexts attached to that error

Alt(Vec<Self>)

A series of parsers were tried at the same location (for instance, via the alt combinator) and all of them failed. All of the errors in this set are โsiblingsโ.

Implementations

impl<I, T, C, E> GenericErrorTree<I, T, C, E>

pub fn map_locations<I2>( self, convert_location: impl FnMut(I) -> I2, ) -> GenericErrorTree<I2, T, C, E>

Convert all of the locations in this error using some kind of mapping function. This is intended to help add additional context that may not have been available when the nom parsers were running, such as line and column numbers.