Type Definition nom_supreme::error::ErrorTree

pub type ErrorTree<I> = GenericErrorTree<I, &'static str, &'static str, Box<dyn Error + Send + Sync + 'static>>;

A comprehensive tree of nom errors describing a parse failure.

This Error type is designed to be VerboseError++. While VerboseError can represent a stack of errors, this type can represent a full tree. In addition to representing a particular specific parse error, it can also represent a stack of nested error contexts (for instance, as provided by context), or a list of alternatives that were all tried individually by alt and all failed.

In general, the design goal for this type is to discard as little useful information as possible. That being said, many ErrorKind variants add very little useful contextual information to error traces; for example, ErrorKind::Alt doesn’t add any interesting context to an ErrorTree::Alt, and its presence in a stack precludes merging together adjacent sets of ErrorTree::Alt siblings.

Examples

Base parser errors

An ErrorTree::Base is an error that occurred at the “bottom” of the stack, from a parser looking for 1 specific kind of thing.

use cool_asserts::assert_matches;
use nom::{Parser, Err};
use nom::character::complete::{digit1, char};
use nom_supreme::error::{ErrorTree, BaseErrorKind, StackContext, Expectation};
use nom_supreme::parser_ext::ParserExt;

let err: Err<ErrorTree<&str>> = digit1.parse("abc").unwrap_err();

assert_matches!(err, Err::Error(ErrorTree::Base{
    location: "abc",
    kind: BaseErrorKind::Expected(Expectation::Digit),
}));

let err: Err<ErrorTree<&str>> = char('a').and(char('b')).parse("acb").unwrap_err();

assert_matches!(err, Err::Error(ErrorTree::Base{
    location: "cb",
    kind: BaseErrorKind::Expected(Expectation::Char('b')),
}));

Stacks

An ErrorTree::Stack is created when a parser combinator—typically context—attaches additional error context to a subparser error. It can have any ErrorTree at the base of the stack.

use cool_asserts::assert_matches;
use nom::{Parser, Err};
use nom::character::complete::{alpha1, space1, char,};
use nom::sequence::{separated_pair, delimited};
use nom_supreme::parser_ext::ParserExt;
use nom_supreme::error::{ErrorTree, BaseErrorKind, StackContext, Expectation};

// Parse a single identifier, defined as just a string of letters.
let identifier = alpha1.context("identifier");

// Parse a pair of identifiers, separated by whitespace
let identifier_pair = separated_pair(identifier, space1, identifier)
    .context("identifier pair");

// Parse a pair of identifiers in parenthesis.
let mut parenthesized = delimited(char('('), identifier_pair, char(')'))
    .context("parenthesized");

let err: Err<ErrorTree<&str>> = parenthesized.parse("(abc 123)").unwrap_err();

assert_matches!(err, Err::Error(ErrorTree::Stack {
    base,
    contexts,
}) => {
    assert_matches!(*base, ErrorTree::Base {
        location: "123)",
        kind: BaseErrorKind::Expected(Expectation::Alpha)
    });

    assert_eq!(contexts, [
        ("123)", StackContext::Context("identifier")),
        ("abc 123)", StackContext::Context("identifier pair")),
        ("(abc 123)", StackContext::Context("parenthesized")),
    ]);
});

Alternatives

An ErrorTree::Alt is created when a series of parsers are all tried, and all of them fail. Most commonly this will happen via the alt combinator or the equivalent .or postfix combinator. When all of these subparsers fail, their errors (each individually their own ErrorTree) are aggregated into an ErrorTree::Alt, indicating that “any one of these things were expected.”

use cool_asserts::assert_matches;
use nom::{Parser, Err};
use nom::branch::alt;
use nom_supreme::error::{ErrorTree, BaseErrorKind, StackContext, Expectation};
use nom_supreme::parser_ext::ParserExt;
use nom_supreme::tag::complete::tag;

let parse_bool = alt((
    tag("true").value(true),
    tag("false").value(false),
));

let mut parse_null_bool = alt((
    parse_bool.map(Some),
    tag("null").value(None),
));

assert_eq!(parse_null_bool.parse("true").unwrap(), ("", Some(true)));
assert_eq!(parse_null_bool.parse("false").unwrap(), ("", Some(false)));
assert_eq!(parse_null_bool.parse("null").unwrap(), ("", None));

let err: Err<ErrorTree<&str>> = parse_null_bool.parse("123").unwrap_err();

// This error communicates to the caller that any one of "true", "false",
// or "null" was expected at that location.
assert_matches!(err, Err::Error(ErrorTree::Alt(choices)) => {
    assert_matches!(choices.as_slice(), [
        ErrorTree::Base {
            location: "123",
            kind: BaseErrorKind::Expected(Expectation::Tag("true"))},
        ErrorTree::Base {
            location: "123",
            kind: BaseErrorKind::Expected(Expectation::Tag("false"))},
        ErrorTree::Base {
            location: "123",
            kind: BaseErrorKind::Expected(Expectation::Tag("null"))},
    ])
});

Contexts and Alternatives

Because Stack and Alt recursively contain ErrorTree errors from subparsers, they can be can combined to create error trees of arbitrary complexity.

use cool_asserts::assert_matches;
use nom::{Parser, Err};
use nom::branch::alt;
use nom_supreme::error::{ErrorTree, BaseErrorKind, StackContext, Expectation};
use nom_supreme::parser_ext::ParserExt;
use nom_supreme::tag::complete::tag;

let parse_bool = alt((
    tag("true").value(true),
    tag("false").value(false),
)).context("bool");

let mut parse_null_bool = alt((
    parse_bool.map(Some),
    tag("null").value(None).context("null"),
)).context("null or bool");

assert_eq!(parse_null_bool.parse("true").unwrap(), ("", Some(true)));
assert_eq!(parse_null_bool.parse("false").unwrap(), ("", Some(false)));
assert_eq!(parse_null_bool.parse("null").unwrap(), ("", None));

let err: Err<ErrorTree<&str>> = parse_null_bool.parse("123").unwrap_err();

assert_matches!(err, Err::Error(ErrorTree::Stack{base, contexts}) => {
    assert_eq!(contexts, [("123", StackContext::Context("null or bool"))]);
    assert_matches!(*base, ErrorTree::Alt(choices) => {
        assert_matches!(&choices[0], ErrorTree::Stack{base, contexts} => {
            assert_eq!(contexts, &[("123", StackContext::Context("bool"))]);
            assert_matches!(&**base, ErrorTree::Alt(choices) => {
                assert_matches!(&choices[0], ErrorTree::Base {
                    location: "123",
                    kind: BaseErrorKind::Expected(Expectation::Tag("true"))
                });
                assert_matches!(&choices[1], ErrorTree::Base {
                    location: "123",
                    kind: BaseErrorKind::Expected(Expectation::Tag("false"))
                });
           });
        });
        assert_matches!(&choices[1], ErrorTree::Stack{base, contexts} => {
            assert_eq!(contexts, &[("123", StackContext::Context("null"))]);
            assert_matches!(&**base, ErrorTree::Base {
                location: "123",
                kind: BaseErrorKind::Expected(Expectation::Tag("null"))
            });
        });
    });
});

Display formatting

TODO WRITE THIS SECTION