Module nom_supreme::tag

Enhanced tag parser for nom.

This module introduces a new trait, TagError, and variants of tag and tag_case_insensitive that can use the trait to add to an error the specific tag that was expected. This allows the error message to report something like Expected tag: "true" instead of just Error: Tag.

Modules

complete

Complete input version of enhanced tag parsers

streaming

Streaming version of enhanced tag parsers.

Traits

TagError

Similar to FromExternalError and ContextError, this trait allows a parser to create an error representing an unmatched tag. This allows error messages to produce more useful context about what went wrong.